+++ Src/InstalledVersion.cs
@@ -1,0 +1,22 @@
+﻿namespace VsixUtil
+{
+    internal class InstalledVersion
+    {
+        internal InstalledVersion(string applicationPath, Version version)
+        {
+            ApplicationPath = applicationPath;
+            Version = version;
+        }
+
+        public Version Version
+        {
+            get;
+        }
+
+        public string ApplicationPath
+        {
+            get;
+        }
+
+    }
+}
