+++ Src/ToolAction.cs
@@ -1,0 +1,10 @@
+﻿namespace VsixUtil
+{
+    internal enum ToolAction
+    {
+        Install,
+        Uninstall,
+        List,
+        Help
+    }
+}
