+++ Src/Version.cs
@@ -1,0 +1,11 @@
+﻿namespace VsixUtil
+{
+    internal enum Version
+    {
+        Vs2010,
+        Vs2012,
+        Vs2013,
+        Vs2015,
+        Vs2017
+    }
+}
