+++ Src/VsixUtil.csproj
@@ -45,7 +45,17 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CommandLine.cs" />
+    <Compile Include="CommandRunner.cs" />
+    <Compile Include="CommonUtil.cs" />
+    <Compile Include="InstalledVersion.cs" />
+    <Compile Include="IVersionManager.cs" />
+    <Compile Include="ProbingPathResolver.cs" />
     <Compile Include="Program.cs" />
+    <Compile Include="ToolAction.cs" />
+    <Compile Include="Version.cs" />
+    <Compile Include="VersionManager.cs" />
+    <Compile Include="VersionManagerRunner.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
